% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{applyExistingstrokeModels}
\alias{applyExistingstrokeModels}
\title{Applies the five existing stroke prediction models}
\usage{
applyExistingstrokeModels(connectionDetails, cdmDatabaseSchema,
  cohortDatabaseSchema, cohortTable, targetId, outcomeId)
}
\arguments{
\item{cohortTable}{The name of the cohort table}

\item{targetId}{The cohort definition id of the target population}

\item{outcomeId}{The cohort definition id of the outcome}

\item{connectioDetails}{The connections details for connecting to the CDM}

\item{cdmDatabaseschema}{The schema holding the CDM data}

\item{cohortDatabaseschema}{The schema holding the cohort table}
}
\value{
A list with the performance and plots
}
\description{
Applies the five existing stroke prediction models
}
\details{
This will run and evaluate five existing stroke risk prediction models
}
