% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getTable1}
\alias{getTable1}
\title{Creates the target population and outcome summary characteristics}
\usage{
getTable1(connectionDetails, cdmDatabaseschema, cohortDatabaseschema,
  cohortTable, targetId, outcomeId, tempCohortTable = "#temp_cohort")
}
\arguments{
\item{cdmDatabaseschema}{The schema holding the CDM data}

\item{cohortDatabaseschema}{The schema holding the cohort table}

\item{cohortTable}{The name of the cohort table}

\item{targetId}{The cohort definition id of the target population}

\item{outcomeId}{The cohort definition id of the outcome}

\item{tempCohortTable}{The name of the temporary table used to hold the cohort}

\item{connectioDetails}{The connections details for connecting to the CDM}
}
\value{
A dataframe with the characteristics
}
\description{
Creates the target population and outcome summary characteristics
}
\details{
This will create the patient characteristic table
}
