% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{packageResults}
\alias{packageResults}
\title{Package the results for sharing with OHDSI researchers}
\usage{
packageResults(outputFolder, dbName, minCellCount = 5)
}
\arguments{
\item{outputFolder}{Name of folder containing the study analysis results}

\item{dbName}{A shareable name for the database used in this study}

\item{minCellCount}{The minimum number of subjects contributing to a count before it can be included in the results.}
}
\description{
Package the results for sharing with OHDSI researchers
}
\details{
This function packages the results.
}
