% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{viewModels}
\alias{viewModels}
\title{View the coefficients of the models in this study and the concept ids used to define them}
\usage{
viewModels()
}
\value{
A data.frame of the models
}
\description{
View the coefficients of the models in this study and the concept ids used to define them
}
\details{
This will print the models and return a data.frame with the models
}
